refactor(frontend): kill phase2Api separate RTK instance (CD-3)
phase2Api was a separate createApi() instance with its own baseQuery,
cache, and tagTypes. It duplicated endpoints from citationApi and
discussionApi, causing stale data issues when one slice invalidated
and the other didn't see it.

Changes:
- Migrate 16 endpoints from phase2Api into existing slices:
  - citationApi: +downloadExport endpoint
  - discussionApi: +listAllDiscussions, +updateMessage, +deleteMessage,
    +getActivityLog, +getActivitySummary, +getEntityActivity, +exportActivityLog
- Unwrap getDiscussion response to match component expectations
- Extend DiscussionThread type with author, participants, messages fields
- Update 12 component imports to use canonical slices (with aliases
  for backward compat: useListAllDiscussionsQuery as useGetDiscussionThreadsQuery,
  useGetDiscussionQuery as useGetDiscussionThreadQuery, etc.)
- Add ActivityLog tagType to apiSlice
- Remove phase2Api from store.ts (reducer + middleware)
- Delete apps/frontend/src/redux/api/phase2Api.ts
- Fix DiscussionThreadCard to import DiscussionThread from slice
- Add null-safety guards for messages/participants in 3 DiscussionDetailClient files

All slices now share one apiSlice instance with consistent cache invalidation.

Phase 8.1.C: Architecture Normalization - Critical Fix
